moved all fetches up to not repeat function code for every broker
